ENH Use (pos,tok) as token
We now save the original (file) position of every token in the token
stream. This is mentioned in the Parsec docs and is a very simple
solution to the problem.

Again, we rely on the fact that the scripts will be small in order to be
able to maintain the whole decorated token stream in memory.