ENH Accept STDIN/STDOUT as a constant
Right now, STDIN is implemented by replacing it with "/dev/stdin" and
STDOUT used "/dev/stdout".

This technique is valid on Unix (including, from the online
documentation, Mac OS X), but I'm not sure it will work on Windows.

There is a check that STDIN can only be used once in a script.