ENH Implement modules
Currently, only ``as_reads`` is implemented as its own, always loaded,
module as a demonstration. However, it may be a good idea to implement
certain builtins as modules to keep the code clean as a possible code
example.

The metadata from the builtin functions was also converted to use the
module infracture. This simplifies the use of the infrastructure as we
now simply have one code path.