ENH Make fqStatC faster
On the benchmark, the time went from 175ms to 18ms, which is even
slightly faster than the lazy IO variant (which clocked in at 21ms).

The resulting code is ugly and very C-like (without C's nicer syntax for
arrays). However, this uglyness is well contained inside fqStatsC, so it
is a reasonable tradeoff.

Suprisingly, a large speed gain was derived from using fromEnum as the
conversion from Word8 to Int! The generic convert performs completely
unnecessary bounds checking, which would kill our inner loop.