ENH Improve speed
When loading Sam header, first just collect all the processed data; then
sort it in parallel (which is faster in the end, and sorting can be
parallelized easily).

Use a hand-written readSamGroups which uses asyncMapEitherC to run
readSamLine in separate threads. This implies that the sam grouping
needs to be fixed in the backend as the last group of a previous element
need not align with the vector boundary