ENH Allow more general language
This introduces temporary variables for every function/operator calls.
For example:

    write(as_reads(samfile('input.sam')), ofile='output.fq')

gets transformed into:

    temp$0 = samfile('input.sam')
    temp$1 = as_reads(temp$0)
    write(temp$1, ofile='output.fq')

The transformation is still incomplete, though, as blocks are untouched.

This makes interpretation easier as it is now more regular. Other
transformation passes are also simplified (in fact, this triggered a few
issues with ``write()`` which are also fixed by this commit).