BUG Propagate exceptions to main thread.
This is still not a perfect solution in every case, but the previous
behaviour would lead to a mysterious "thread blocked indefinitely in an
STM transaction" on errors. At least now, the user gets an intelligible
error message.

This is a better solution to an old issue which would cause a crash.
Now, we propagate the exception.