ENH Add split/align for NGLess
This automates the functionality that MOCAT required the user to provide
by hand (namely splitting large databases so they could be used). The
user needs to provide the size to split in.

WIP: Still a bit rough in terms of producing the right output files and
being robust to failures, only partially, but likely working.