BUG Fix edge case in select()
Previously, it was possible that the samlines with the sequence would be
filtered out of the SAM/BAM, while lines without sequence information
would be kept.

As this causes a change in behaviour, it triggers an update in version
header. However, the case where behaviour changes is rare enough that
the old behaviour is retained when a version prior to 0.8 is requested.
A warning is emitted when select() is used with older version
declarations.