RFCT Separate filename matching from loading
Makes testing much easier too: we can do it using unit tests instead of
integration tests. This is much faster and potentially scales much
better (just add 1 line of code to add a new test rather than a full
directory of files).

For now, kept some of the older files that also test other parts of the
system, though. In particular, `mocat_sample_bz2_paired_mixed` shows how
the results change in a way that produces equivalent, but not identical
results: the order of the reads in the FQ file changes.